+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.8.1
+- Added: Blacktalon Quartermaster not selling legendary cloak bases. Please fix
+- Added: Travel to Val and Naigtal to Quell Leaders of the Void
+
 ## 1.8.0
 - Added: Addon Compartment launcher support.
 - Updated: Minimap button visibility with Minimap Button Bag.
@@ -327,6 +331,7 @@
 - Added external link copy and guild chat sharing for selected posts.
 - Added login toasts for recent posts.
 - Added class section shortcuts for class hotfix posts.
+